frontend/requirements.txt
gradio
grpcio
requests
